# DataServer/api/openapi.yaml
...
components:
    parameters:
        AuthorizationHeader:
            description: |
                Bearer authorization header. Depending on the route, the
                token is either the configured `VELOX_ADMIN_TOKEN` or an M2M
                credential validated by the route's middleware. Example:
                `Authorization: Bearer eyJhbGciOi...`.
            in: header
            name: Authorization
            required: true
            schema:
                example: Bearer ${VELOX_ADMIN_TOKEN}
                pattern: ^Bearer\s+.+
                type: string
...
        XRequestIDHeader:
            description: |
                Optional client-supplied correlation ID, forwarded verbatim
                into structured observability logs as the request trace key.
                Convention: short, prefixed by the caller's domain (e.g.
                `creator-push-smoke`, `e2e-<runner-host>-<ts>`). The Master
                does NOT validate the format; an absent header is replaced
                server-side with a generated UUID.
            in: header
            name: X-Request-ID
            required: false
            schema:
                example: creator-push-smoke
                maxLength: 128
                type: string
    schemas:
...
        ErrorEnvelope:
            description: |
                Single error envelope returned for 4xx/5xx responses.
                Stable shape across the pipeline group; for 422 the
                `details` array identifies which top-level or nested fields
                failed validation.
            properties:
                details:
                    description: Field-level failure list (only present for 422).
                    items:
                        properties:
                            issue:
                                example: required
                                type: string
                            path:
                                example: payload.scenes.0.duration_seconds
                                type: string
                        type: object
                    type: array
                error:
                    $ref: '#/components/schemas/ErrorCode'
                message:
                    description: Human-readable explanation.
                    example: payload.scenes[0] missing 'duration_seconds'
                    type: string
                ok:
                    enum:
                        - false
                    example: false
                    type: boolean
            required:
                - ok
                - error
                - message
            type: object
        ExternalAssemblyRequest:
            properties:
                assets:
                    items:
                        properties:
                            asset_id:
                                type: string
                            availability:
                                type: string
                            kind:
                                type: string
                            mime_type:
                                type: string
                            producer:
                                type: string
                            profile_id:
                                type: string
                            required:
                                type: boolean
                            sha256:
                                type: string
                            size_bytes:
                                type: integer
                            state:
                                type: string
                            url:
                                type: string
                        type: object
                    minItems: 0
                    type: array
                early_manifest:
                    properties:
                        assets:
                            items:
                                properties:
                                    asset_id:
                                        type: string
                                    availability:
                                        type: string
                                    kind:
                                        type: string
                                    mime_type:
                                        type: string
                                    producer:
                                        type: string
                                    profile_id:
                                        type: string
                                    required:
                                        type: boolean
                                    sha256:
                                        type: string
                                    size_bytes:
                                        type: integer
                                    state:
                                        type: string
                                    url:
                                        type: string
                                type: object
                            minItems: 0
                            type: array
                        contract_version:
                            type: string
                        dispatch:
                            properties:
                                mode:
                                    type: string
                                start_prepare:
                                    type: boolean
                                target:
                                    type: string
                            type: object
                        expected_profile:
                            type: string
                        job_id:
                            type: string
                        preparation_hash:
                            type: string
                        profile:
                            properties:
                                b_frames:
                                    type: integer
                                closed_gop:
                                    type: boolean
                                codec:
                                    type: string
                                codec_level:
                                    type: string
                                codec_profile:
                                    type: string
                                container:
                                    type: string
                                container_layout:
                                    type: string
                                fps_den:
                                    type: integer
                                fps_num:
                                    type: integer
                                gop_size:
                                    type: integer
                                height:
                                    type: integer
                                pixel_format:
                                    type: string
                                profile_id:
                                    type: string
                                time_base_den:
                                    type: integer
                                time_base_num:
                                    type: integer
                                version:
                                    type: integer
                                width:
                                    type: integer
                            type: object
                        revision:
                            type: integer
                        timeline:
                            items:
                                properties:
                                    asset_id:
                                        type: string
                                    scene_id:
                                        type: string
                                type: object
                            minItems: 0
                            type: array
                        timeline_revision:
                            type: integer
                    type: object
                output_profile:
                    type: string
                send_to_velox:
                    type: boolean
                timeline:
                    items:
                        properties:
                            asset_id:
                                type: string
                            scene_id:
                                type: string
                        type: object
                    minItems: 0
                    type: array
                timeline_hash:
                    type: string
                timeline_revision:
                    type: integer
            type: object
...
        SubmitDeliveryPlanEntry:
            properties:
                destination_id:
                    minLength: 1
                    type: string
                metadata:
                    additionalProperties: {}
                    type: object
                priority:
                    minimum: 0
                    type: integer
                retry_budget:
                    minimum: 0
                    type: integer
            required:
                - destination_id
            type: object
        SubmitJobAcceptedResponse:
            description: |
                Successful creation via the simplified job submission endpoint.
                The handler has persisted forwarding + Job + Task atomically
                via creatorflow.Resolver. The `status_url` field carries
                the canonical polling path for this job_id, and the response
                ALSO emits `Location: /api/v1/jobs/{job_id}` per RFC 7231
                so RESTful clients can chain immediately via the header
                without parsing the body.
            properties:
                accepted_from:
                    description: Fixed marker identifying this intake path.
                    enum:
                        - api_v1_jobs
                    example: api_v1_jobs
                    type: string
                dispatch_status:
                    enum:
                        - queued_for_workers
                    example: queued_for_workers
                    type: string
                idempotency_key:
                    description: Echoed from the request.
                    example: ext-job-20260727-001
                    type: string
                job_id:
                    description: |
                        Velox-side job identifier for downstream polling.
                        Mirrored in the `Location` response header.
                    example: job_a1b2c3d4
                    type: string
                ok:
                    enum:
                        - true
                    example: true
                    type: boolean
                status_url:
                    description: |
                        Canonical polling URL for this job_id. Env-relative
                        (no scheme/host) so it stays stable across dev /
                        staging / production environments and across API
                        gateway rewrites of the Location header. Clients
                        can dereference this with the same auth headers as
                        the POST.
                    example: /api/v1/jobs/job_a1b2c3d4
                    type: string
            required:
                - ok
                - accepted_from
                - idempotency_key
                - job_id
                - dispatch_status
                - status_url
            type: object
...
        SubmitJobRequest:
            properties:
                assembly:
                    $ref: '#/components/schemas/ExternalAssemblyRequest'
                compiled_render_plan_json:
                    type: string
                compiled_render_plan_sha256:
                    maxLength: 64
                    minLength: 64
                    type: string
                delivery_plan:
                    items:
                        $ref: '#/components/schemas/SubmitDeliveryPlanEntry'
                    minItems: 0
                    type: array
                idempotency_key:
                    maxLength: 128
                    minLength: 1
                    type: string
                job_type:
                    type: string
                layers:
                    items:
                        $ref: '#/components/schemas/SubmitLayer'
                    minItems: 0
                    type: array
                manifest_ref:
                    $ref: '#/components/schemas/SubmitManifestRef'
                output:
                    properties:
                        format:
                            enum:
                                - mp4
                                - mov
                                - webm
                            type: string
                        fps:
                            maximum: 240
                            minimum: 1
                            type: integer
                        height:
                            maximum: 16384
                            minimum: 1
                            type: integer
                        width:
                            maximum: 16384
                            minimum: 1
                            type: integer
                    type: object
                overlays:
                    items:
                        properties:
                            asset_id:
                                maxLength: 512
                                type: string
                            audio_mode:
                                enum:
                                    - preserve_final_audio
                                type: string
                            drive_file_id:
                                maxLength: 128
                                type: string
                            end_frame:
                                minimum: 1
                                type: integer
                            frame_count:
                                minimum: 1
                                type: integer
                            id:
                                maxLength: 128
                                type: string
                            mode:
                                enum:
                                    - replace
                                    - composite
                                type: string
                            sha256:
                                type: string
                            start_frame:
                                minimum: 0
                                type: integer
                            url:
                                type: string
                            z_index:
                                type: integer
                        required:
                            - audio_mode
                            - id
                            - mode
                        type: object
                    maxItems: 10000
                    minItems: 0
                    type: array
                placement_pin_worker_id:
                    maxLength: 128
                    type: string
                publications:
                    items:
                        properties:
                            default_language:
                                type: string
                            destinations:
                                items:
                                    properties:
                                        destination_id:
                                            minLength: 1
                                            type: string
                                        metadata_override:
                                            properties:
                                                category_id:
                                                    type: string
                                                contains_synthetic_media:
                                                    type: boolean
                                                description:
                                                    type: string
                                                made_for_kids:
                                                    type: boolean
                                                privacy:
                                                    type: string
                                                publish_at:
                                                    type: string
                                                tags:
                                                    items:
                                                        type: string
                                                    minItems: 0
                                                    type: array
                                                title:
                                                    type: string
                                            type: object
                                        priority:
                                            minimum: 0
                                            type: integer
                                        provider_options:
                                            additionalProperties: {}
                                            type: object
                                        retry_budget:
                                            minimum: 0
                                            type: integer
                                    required:
                                        - destination_id
                                    type: object
                                minItems: 0
                                minimum: 1
                                type: array
                            language:
                                type: string
                            localizations:
                                additionalProperties:
                                    properties:
                                        description:
                                            type: string
                                        title:
                                            type: string
                                    type: object
                                type: object
                            metadata:
                                properties:
                                    category_id:
                                        type: string
                                    contains_synthetic_media:
                                        type: boolean
                                    description:
                                        type: string
                                    made_for_kids:
                                        type: boolean
                                    privacy:
                                        type: string
                                    publish_at:
                                        type: string
                                    tags:
                                        items:
                                            type: string
                                        minItems: 0
                                        type: array
                                    title:
                                        type: string
                                type: object
                            output_ref:
                                properties:
                                    artifact_role:
                                        type: string
                                    variant_id:
                                        type: string
                                type: object
                            provider_options:
                                additionalProperties: {}
                                type: object
                            publication_id:
                                minLength: 1
                                type: string
                        required:
                            - destinations
                            - output_ref
                            - publication_id
                        type: object
                    minItems: 0
                    type: array
                publishing_target:
                    $ref: '#/components/schemas/SubmitPublishingTarget'
                scenes:
                    items:
                        $ref: '#/components/schemas/SubmitScene'
                    maxItems: 10000
                    minItems: 0
                    type: array
                script_text:
                    type: string
                spec:
                    additionalProperties: {}
                    type: object
                template_id:
                    type: string
                template_version:
                    type: integer
                video_name:
                    maxLength: 300
                    type: string
                visual_replacements:
                    items:
                        $ref: '#/components/schemas/SubmitVisualReplacement'
                    maxItems: 10000
                    minItems: 0
                    type: array
            required:
                - idempotency_key
            type: object
...
        SubmitLayer:
            properties:
                animation:
                    type: string
                asset:
                    type: string
                duration_seconds:
                    format: float
                    minimum: 0
                    type: number
                font:
                    type: string
                font_size:
                    format: float
                    minimum: 0
                    type: number
                id:
                    minLength: 1
                    type: string
                position:
                    items:
                        format: float
                        type: number
                    maxItems: 2
                    minItems: 2
                    type: array
                preset:
                    type: string
                role:
                    enum:
                        - title
                        - name
                        - important_phrase
                        - overlay
                    type: string
                source:
                    type: string
                start_seconds:
                    format: float
                    minimum: 0
                    type: number
                text:
                    type: string
                type:
                    enum:
                        - text
                        - image
                        - video
                        - color
                    type: string
            required:
                - id
                - type
            type: object
        SubmitManifestRef:
            properties:
                schema_version:
                    enum:
                        - velox.render-manifest.v1
                    type: string
                sha256:
                    maxLength: 64
                    minLength: 64
                    pattern: ^[0-9a-f]{64}$
                    type: string
                url:
                    maxLength: 2048
                    minLength: 1
                    pattern: ^(https?://|velox-asset://).+
                    type: string
            required:
                - schema_version
                - sha256
                - url
            type: object
...
    securitySchemes:
        bearerAdminToken:
            bearerFormat: opaque
            description: |
                Bearer token matching `cfg.Auth.AdminToken`, which is loaded
                from the `VELOX_ADMIN_TOKEN` env var on the Master process
                (see `DataServer/internal/config/config_misc.go::loadAuth`).
                The token is opaque to clients; rotate via `scripts/rotate_token.sh`
                + restart. Minimum length is enforced server-side. A missing
                or mismatched bearer returns `401 Unauthorized`; never echo
                the token value in client logs.
            scheme: bearer
            type: http
        bearerM2MToken:
            bearerFormat: opaque
            description: |
                M2M bearer issued by the Velox administrator. Job routes
                require the `jobs.submit` scope and are validated by the
                M2M key middleware.
            scheme: bearer
            type: http
info:
    contact:
        name: Velox architecture
        url: https://github.com/example/velox
    description: |
        The Velox Master exposes a small, authenticated HTTP API for
        programmatic intake of creator-built payloads and for orchestrator
        control. API endpoints are protected by bearer authorization. Most
        legacy routes use the configured `VELOX_ADMIN_TOKEN`; M2M publishing
        and job-submission routes use the scoped credential middleware and
        require the `jobs.submit` scope.

        This revision focuses on the new **creator_push** intake path
        (`POST /api/v1/creator/jobs`) introduced alongside the existing
        async `CreatorForwardingRunner` poller. Both intake paths converge
        on the canonical `creatorflow.Resolver` (the unique writer of
        `creator_forwardings`/`jobs`/`tasks`); the push handler does NOT
        open its own transaction or duplicate schema. See
        `docs/architecture/current-architecture.md` §12 and §4.2 of
        `docs/architecture/runtime-invariants.md` for the invariant.

        The HTTP contract is also documented narratively in
        `docs/CREATOR-PUSH.md`. Whenever the two disagree, the tested
        behavior in `DataServer/internal/handlers/server/pipeline/creator_push_e2e_test.go`
        is the source of truth.

        NOTE ON SCHEMA REDUNDANCY: this spec declares BOTH `CreatorPushPayload`
        (the flat wire-facing schema the Creator machine sends) and
...
openapi: 3.1.0
paths:
...
    /api/v1/jobs/pre:
        post:
            operationId: prepareJob
            parameters:
                - $ref: '#/components/parameters/AuthorizationHeader'
                - $ref: '#/components/parameters/XRequestIDHeader'
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SubmitJobRequest'
                required: true
            responses:
                "202":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SubmitJobAcceptedResponse'
                    description: Accepted.
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request.
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized.
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Conflict.
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unprocessable Entity.
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error.
            security:
                - bearerM2MToken: []
            summary: Create one job and prefetch its initial assets before runtime generation.
            tags:
                - jobs
...
security:
    - bearerAdminToken: []
servers:
...